Refactor server models
